drainer: move lifecycleHookCache out of a global var
